+++ modules/holonews/api.php
@@ -1,0 +1,3 @@
+<?php
+Route::get('posts', 'ArticlesController@all')->name('posts.list');
+Route::get('posts/{slug}', 'ArticlesController@get')->name('posts.show');
